# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'My API Documentation'
  description: 'API documentation for my Laravel application'
  contact:
    email: admin@example.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'API Server'
paths:
...
  /api/cart:
    get:
      tags:
        - Cart
      summary: 'Get all cart items for the authenticated user'
      operationId: getCart
      responses:
        200:
          description: 'Cart items retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        user_id:
                          type: integer
                          example: 1
                        product_id:
                          type: integer
                          example: 3
                        quantity:
                          type: integer
                          example: 2
                        product:
                          properties:
                            id:
                              type: integer
                              example: 3
                            name:
                              type: string
                              example: 'Rose Perfume'
                            price:
                              type: number
                              format: float
                              example: 29.99
                            image:
                              type: string
                              example: products/abc.jpg
                              nullable: true
                          type: object
                      type: object
                  total:
                    description: 'Sum of (product price × quantity) across all cart items'
                    type: number
                    format: float
                    example: 59.98
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
      security:
        -
          sanctum: []
    post:
      tags:
        - Cart
      summary: 'Add a product to cart — increments quantity if product already exists in cart'
      operationId: addToCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - product_id
              properties:
                product_id:
                  description: 'Required — must exist in products table'
                  type: integer
                  example: 3
                quantity:
                  description: 'Optional — must be integer min 1, defaults to 1 if omitted'
                  type: integer
                  example: 2
                  minimum: 1
              type: object
      responses:
        201:
          description: 'Product added to cart (or quantity incremented if already in cart)'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      user_id:
                        type: integer
                        example: 1
                      product_id:
                        type: integer
                        example: 3
                      quantity:
                        type: integer
                        example: 2
                      product:
                        properties:
                          id:
                            type: integer
                            example: 3
                          name:
                            type: string
                            example: 'Rose Perfume'
                          price:
                            type: number
                            format: float
                            example: 29.99
                          image:
                            type: string
                            example: products/abc.jpg
                            nullable: true
                        type: object
                    type: object
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
        422:
          description: 'Validation failed'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The product_id field is required.'
                  errors:
                    properties:
                      product_id:
                        type: array
                        items:
                          type: string
                          example: 'The selected product_id is invalid.'
                      quantity:
                        type: array
                        items:
                          type: string
                          example: 'The quantity must be at least 1.'
                    type: object
                type: object
      security:
        -
          sanctum: []
  '/api/cart/{id}':
    put:
      tags:
        - Cart
      summary: 'Update the quantity of a specific cart item'
      operationId: updateCart
      parameters:
        -
          name: id
          in: path
          description: 'Cart item ID — must belong to the authenticated user'
          required: true
          schema:
            type: integer
            example: 1
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - quantity
              properties:
                quantity:
                  description: 'Required — integer min 1, replaces the current quantity'
                  type: integer
                  example: 5
                  minimum: 1
              type: object
      responses:
        200:
          description: 'Cart item quantity updated successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      user_id:
                        type: integer
                        example: 1
                      product_id:
                        type: integer
                        example: 3
                      quantity:
                        type: integer
                        example: 5
                      product:
                        properties:
                          id:
                            type: integer
                            example: 3
                          name:
                            type: string
                            example: 'Rose Perfume'
                          price:
                            type: number
                            format: float
                            example: 29.99
                          image:
                            type: string
                            example: products/abc.jpg
                            nullable: true
                        type: object
                    type: object
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
        404:
          description: 'Cart item not found or does not belong to the authenticated user'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Not found'
                type: object
        422:
          description: 'Validation failed'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The quantity field is required.'
                  errors:
                    properties:
                      quantity:
                        type: array
                        items:
                          type: string
                          example: 'The quantity must be at least 1.'
                    type: object
                type: object
      security:
        -
          sanctum: []
    delete:
      tags:
        - Cart
      summary: 'Remove a specific item from the cart'
      operationId: removeFromCart
      parameters:
        -
          name: id
          in: path
          description: 'Cart item ID — must belong to the authenticated user'
          required: true
          schema:
            type: integer
            example: 1
            minimum: 1
      responses:
        200:
          description: 'Cart item removed successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Item removed'
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
        404:
          description: 'Cart item not found or does not belong to the authenticated user'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Not found'
                type: object
      security:
        -
          sanctum: []
...
components:
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Enter token in format: Bearer <token>'
      bearerFormat: JWT
      scheme: bearer
...